+++ .internal/assemblyscript/@gl/api/w2h/pickup.ts
@@ -1,5 +1,7 @@
 export declare function offerPickup(slug: string): void;
-export declare function get(tags: string[]): string[];
+export declare function query(tag: string): boolean;
+export declare function toggle(slug: string, value: boolean): void;
 
 export const _keep_offerPickup = offerPickup;
-export const _keep_get = get;
+export const _keep_query = query;
+export const _keep_toggle = toggle;