+++ .internal/assemblyscript/@gl/api/w2h/sound.ts
@@ -4,7 +4,7 @@ import { LoadOpts, PlayOpts } from "../types/sound";
  * Loads a sound and returns its ID.
  *
  * @param opts Sound options.
- * @returns The ID of the loaded sound.
+ * @returns The *asset* ID of the loaded sound.
  */
 export declare function loadSound(opts: LoadOpts): i32;
 export declare function playSound(opts: PlayOpts): i32;
@@ -14,8 +14,14 @@ export declare function setVolume(
   soundId: i32,
   volume: f32
 ): void;
+export declare function crossfade(
+  assetAId: i32,
+  assetBId: i32,
+  duration: i32
+): void;
 
 export const _keep_loadSound = loadSound;
 export const _keep_playSound = playSound;
 export const _keep_stopSound = stopSound;
 export const _keep_setVolume = setVolume;
+export const _keep_crossfade = crossfade;